pyproject.toml
[project]
name = "pg-langgraph-agent"
version = "0.1.0"
description = "PostgreSQL SQL agent with memory using LangGraph"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "langgraph>=1.0.0",
    "langchain>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-anthropic>=1.0.0",
    "langchain-community>=0.4",
    "jupyter>=1.0.0",
    "langgraph-cli[inmem]>=0.4.0",
    "langchain-mcp-adapters",
    "python-dotenv",
    "psycopg2-binary",
    "click",
]
